+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Version 0
 
+### v0.17.0
+
+- Added `handshake` property to the client objects:
+  - See the Socket.IO documentation on handshake: https://socket.io/docs/v4/server-socket-instance/#sockethandshake
+
 ### v0.16.0
 
 - Added `.emit()` method to the clients returned by `getClients()` method of `all` or `withRoom()`;